// frontend/src/types/types.ts
export interface OutbreakData {
  Id: string;
  LastModified: string;
  PublicationDate: string;
  DateCreated: string;
  IncludeInSitemap: boolean;
  SystemSourceKey: string | null;
  UrlName: string;
  ItemDefaultUrl: string;
  Response: string;
  FurtherInformation: string;
  Summary: string;
  PublicationDateAndTime: string;
  TitleSuffix: string;
  UseOverrideTitle: boolean;
  Title: string;
  Epidemiology: string | null;
  OverrideTitle: string;
  Advice: string;
  Assessment: string;
  Overview: string;
  DonId: string;
  Provider: string;
}